Show continuous status information on add --wait
This prints an ASCII process bar on add --wait, which is updated every
second.

This complements the status command. It saves the trouble of polling the
status output yourself and behaves more like traditional download
programs. On the other hand it also means more work for aria, since it
gets more RPC-calls querying for the download status. So there should
probably be an option to turn the new behaviors off, maybe called
--quiet.